add scoverage plugin and refactoring the code
