+++ DBS/src/test/scala/ConnectionTest.scala
@@ -6,8 +6,14 @@ import reactivemongo.bson.BSON
 class ConnectorTest extends FlatSpec with Connector{
 
     "DB connection table name" should "table1" in{    
-    val res=connector.name
+    val res=connector.get.name
     val expectedres="table1"
     assert(res===expectedres)
   }
+    
+    "DB we cant make a connection" should "None" in{
+      val res=connector.get.name
+      val expectedres=None
+      assert(res===expectedres)
+    }
 }
